requirements.txt
tushare>=1.3.0
pyyaml>=6.0